move harmonics to a proper subpackage
