+++ src/eko/anomalous_dimensions/__init__.py
@@ -20,7 +20,8 @@ terms of the anomalous dimensions (note the additional sign!)
 import numba as nb
 import numpy as np
 
-from . import as1, as2, as3, harmonics
+from .. import harmonics
+from . import as1, as2, as3
 
 
 @nb.njit(cache=True)
@@ -102,7 +103,7 @@ def gamma_ns(order, mode, n, nf):
         eko.anomalous_dimensions.as3.gamma_nsv : :math:`\gamma_{ns,v}^{(2)}(N)`
     """
     # cache the s-es
-    sx = np.full(1, harmonics.harmonic_S1(n))
+    sx = np.full(1, harmonics.S1(n))
     # now combine
     gamma_ns = np.zeros(order + 1, np.complex_)
     gamma_ns[0] = as1.gamma_ns(n, sx[0])
@@ -119,8 +120,8 @@ def gamma_ns(order, mode, n, nf):
         gamma_ns[1] = gamma_ns_1
     # NNLO and beyond
     if order >= 2:
-        sx = np.append(sx, harmonics.harmonic_S2(n))
-        sx = np.append(sx, harmonics.harmonic_S3(n))
+        sx = np.append(sx, harmonics.S2(n))
+        sx = np.append(sx, harmonics.S3(n))
         if mode == 10101:
             gamma_ns_2 = -as3.gamma_nsp(n, nf, sx)
         elif mode == 10201:
@@ -157,16 +158,16 @@ def gamma_singlet(order, n, nf):
         eko.anomalous_dimensions.as3.gamma_singlet : :math:`\gamma_{S}^{(2)}(N)`
     """
     # cache the s-es
-    sx = np.full(1, harmonics.harmonic_S1(n))
+    sx = np.full(1, harmonics.S1(n))
     if order >= 1:
-        sx = np.append(sx, harmonics.harmonic_S2(n))
-        sx = np.append(sx, harmonics.harmonic_S3(n))
+        sx = np.append(sx, harmonics.S2(n))
+        sx = np.append(sx, harmonics.S3(n))
 
     gamma_s = np.zeros((order + 1, 2, 2), np.complex_)
     gamma_s[0] = as1.gamma_singlet(n, sx[0], nf)
     if order >= 1:
         gamma_s[1] = as2.gamma_singlet(n, nf)
     if order == 2:
-        sx = np.append(sx, harmonics.harmonic_S4(n))
+        sx = np.append(sx, harmonics.S4(n))
         gamma_s[2] = -as3.gamma_singlet(n, nf, sx)
     return gamma_s
