+++ src/eko/anomalous_dimensions/as2.py
@@ -10,8 +10,8 @@ involving ``FormGet`` :cite:`Hahn:2016ebn`.
 import numba as nb
 import numpy as np
 
-from .. import constants
-from . import harmonics
+from .. import constants, harmonics
+from ..harmonics.constants import log2, zeta2, zeta3
 
 
 @nb.njit(cache=True)
@@ -34,19 +34,18 @@ def gamma_nsm(n, nf):
             |NLO| valence-like non-singlet anomalous dimension
             :math:`\\gamma_{ns,-}^{(1)}(N)`
     """
-    S1 = harmonics.harmonic_S1(n)
-    S2 = harmonics.harmonic_S2(n)
+    # TODO: use harmonics from cache
+    S1 = harmonics.S1(n)
+    S2 = harmonics.S2(n)
     # Here, Sp refers to S' ("s-prime") (german: "s-strich" or in Pegasus language: SSTR)
     # of :cite:`Gluck:1989ze` and NOT to the Spence function a.k.a. dilogarithm
-    Sp1m = harmonics.harmonic_S1((n - 1) / 2)
-    Sp2m = harmonics.harmonic_S2((n - 1) / 2)
-    Sp3m = harmonics.harmonic_S3((n - 1) / 2)
-    g3n = harmonics.mellin_g3(n)
-    zeta2 = harmonics.zeta2
-    zeta3 = harmonics.zeta3
+    Sp1m = harmonics.S1((n - 1) / 2)
+    Sp2m = harmonics.S2((n - 1) / 2)
+    Sp3m = harmonics.S3((n - 1) / 2)
+    g3n = harmonics.g_functions.mellin_g3(n)
     # fmt: off
-    gqq1m_cfca = 16*g3n - (144 + n*(1 + n)*(156 + n*(340 + n*(655 + 51*n*(2 + n)))))/(18.*np.power(n,3)*np.power(1 + n,3)) + (-14.666666666666666 + 8/n - 8/(1 + n))*S2 - (4*Sp2m)/(n + np.power(n,2)) + S1*(29.77777777777778 + 16/np.power(n,2) - 16*S2 + 8*Sp2m) + 2*Sp3m + 10*zeta3 + zeta2*(16*S1 - 16*Sp1m - (16*(1 + n*np.log(2)))/n) # pylint: disable=line-too-long
-    gqq1m_cfcf = -32*g3n + (24 - n*(-32 + 3*n*(-8 + n*(3 + n)*(3 + np.power(n,2)))))/(2.*np.power(n,3)*np.power(1 + n,3)) + (12 - 8/n + 8/(1 + n))*S2 + S1*(-24/np.power(n,2) - 8/np.power(1 + n,2) + 16*S2 - 16*Sp2m) + (8*Sp2m)/(n + np.power(n,2)) - 4*Sp3m - 20*zeta3 + zeta2*(-32*S1 + 32*Sp1m + 32*(1/n + np.log(2))) # pylint: disable=line-too-long
+    gqq1m_cfca = 16*g3n - (144 + n*(1 + n)*(156 + n*(340 + n*(655 + 51*n*(2 + n)))))/(18.*np.power(n,3)*np.power(1 + n,3)) + (-14.666666666666666 + 8/n - 8/(1 + n))*S2 - (4*Sp2m)/(n + np.power(n,2)) + S1*(29.77777777777778 + 16/np.power(n,2) - 16*S2 + 8*Sp2m) + 2*Sp3m + 10*zeta3 + zeta2*(16*S1 - 16*Sp1m - (16*(1 + n*log2))/n) # pylint: disable=line-too-long
+    gqq1m_cfcf = -32*g3n + (24 - n*(-32 + 3*n*(-8 + n*(3 + n)*(3 + np.power(n,2)))))/(2.*np.power(n,3)*np.power(1 + n,3)) + (12 - 8/n + 8/(1 + n))*S2 + S1*(-24/np.power(n,2) - 8/np.power(1 + n,2) + 16*S2 - 16*Sp2m) + (8*Sp2m)/(n + np.power(n,2)) - 4*Sp3m - 20*zeta3 + zeta2*(-32*S1 + 32*Sp1m + 32*(1/n + log2)) # pylint: disable=line-too-long
     gqq1m_cfnf = (-12 + n*(20 + n*(47 + 3*n*(2 + n))))/(9.*np.power(n,2)*np.power(1 + n,2)) - (40*S1)/9. + (8*S2)/3. # pylint: disable=line-too-long
     # fmt: on
     result = constants.CF * (
@@ -77,17 +76,15 @@ def gamma_nsp(n, nf):
             |NLO| singlet-like non-singlet anomalous dimension
             :math:`\\gamma_{ns,+}^{(1)}(N)`
     """
-    S1 = harmonics.harmonic_S1(n)
-    S2 = harmonics.harmonic_S2(n)
-    Sp1p = harmonics.harmonic_S1(n / 2)
-    Sp2p = harmonics.harmonic_S2(n / 2)
-    Sp3p = harmonics.harmonic_S3(n / 2)
-    g3n = harmonics.mellin_g3(n)
-    zeta2 = harmonics.zeta2
-    zeta3 = harmonics.zeta3
+    S1 = harmonics.S1(n)
+    S2 = harmonics.S2(n)
+    Sp1p = harmonics.S1(n / 2)
+    Sp2p = harmonics.S2(n / 2)
+    Sp3p = harmonics.S3(n / 2)
+    g3n = harmonics.g_functions.mellin_g3(n)
     # fmt: off
-    gqq1p_cfca = -16*g3n + (132 - n*(340 + n*(655 + 51*n*(2 + n))))/(18.*np.power(n,2)*np.power(1 + n,2)) + (-14.666666666666666 + 8/n - 8/(1 + n))*S2 - (4*Sp2p)/(n + np.power(n,2)) + S1*(29.77777777777778 - 16/np.power(n,2) - 16*S2 + 8*Sp2p) + 2*Sp3p + 10*zeta3 + zeta2*(16*S1 - 16*Sp1p + 16*(1/n - np.log(2))) # pylint: disable=line-too-long
-    gqq1p_cfcf = 32*g3n - (8 + n*(32 + n*(40 + 3*n*(3 + n)*(3 + np.power(n,2)))))/(2.*np.power(n,3)*np.power(1 + n,3)) + (12 - 8/n + 8/(1 + n))*S2 + S1*(40/np.power(n,2) - 8/np.power(1 + n,2) + 16*S2 - 16*Sp2p) + (8*Sp2p)/(n + np.power(n,2)) - 4*Sp3p - 20*zeta3 + zeta2*(-32*S1 + 32*Sp1p + 32*(-(1/n) + np.log(2))) # pylint: disable=line-too-long
+    gqq1p_cfca = -16*g3n + (132 - n*(340 + n*(655 + 51*n*(2 + n))))/(18.*np.power(n,2)*np.power(1 + n,2)) + (-14.666666666666666 + 8/n - 8/(1 + n))*S2 - (4*Sp2p)/(n + np.power(n,2)) + S1*(29.77777777777778 - 16/np.power(n,2) - 16*S2 + 8*Sp2p) + 2*Sp3p + 10*zeta3 + zeta2*(16*S1 - 16*Sp1p + 16*(1/n - log2)) # pylint: disable=line-too-long
+    gqq1p_cfcf = 32*g3n - (8 + n*(32 + n*(40 + 3*n*(3 + n)*(3 + np.power(n,2)))))/(2.*np.power(n,3)*np.power(1 + n,3)) + (12 - 8/n + 8/(1 + n))*S2 + S1*(40/np.power(n,2) - 8/np.power(1 + n,2) + 16*S2 - 16*Sp2p) + (8*Sp2p)/(n + np.power(n,2)) - 4*Sp3p - 20*zeta3 + zeta2*(-32*S1 + 32*Sp1p + 32*(-(1/n) + log2)) # pylint: disable=line-too-long
     gqq1p_cfnf = (-12 + n*(20 + n*(47 + 3*n*(2 + n))))/(9.*np.power(n,2)*np.power(1 + n,2)) - (40*S1)/9. + (8*S2)/3. # pylint: disable=line-too-long
     # fmt: on
     result = constants.CF * (
@@ -145,9 +142,9 @@ def gamma_qg(n, nf):
             |NLO| quark-gluon singlet anomalous dimension
             :math:`\\gamma_{qg}^{(1)}(N)`
     """
-    S1 = harmonics.harmonic_S1(n)
-    S2 = harmonics.harmonic_S2(n)
-    Sp2p = harmonics.harmonic_S2(n / 2)
+    S1 = harmonics.S1(n)
+    S2 = harmonics.S2(n)
+    Sp2p = harmonics.S2(n / 2)
     # fmt: off
     gqg1_nfca = (-4*(16 + n*(64 + n*(104 + n*(128 + n*(85 + n*(36 + n*(25 + n*(15 + n*(6 + n))))))))))/((-1 + n)*np.power(n,3)*np.power(1 + n,3)*np.power(2 + n,3)) - (16*(3 + 2*n)*S1)/np.power(2 + 3*n + np.power(n,2),2) + (4*(2 + n + np.power(n,2))*np.power(S1,2))/(n*(2 + 3*n + np.power(n,2))) - (4*(2 + n + np.power(n,2))*S2)/(n*(2 + 3*n + np.power(n,2))) + (4*(2 + n + np.power(n,2))*Sp2p)/(n*(2 + 3*n + np.power(n,2))) # pylint: disable=line-too-long
     gqg1_nfcf = (-2*(4 + n*(8 + n*(1 + n)*(25 + n*(26 + 5*n*(2 + n))))))/(np.power(n,3)*np.power(1 + n,3)*(2 + n)) + (8*S1)/np.power(n,2) - (4*(2 + n + np.power(n,2))*np.power(S1,2))/(n*(2 + 3*n + np.power(n,2))) + (4*(2 + n + np.power(n,2))*S2)/(n*(2 + 3*n + np.power(n,2))) # pylint: disable=line-too-long
@@ -178,9 +175,9 @@ def gamma_gq(n, nf):
             |NLO| gluon-quark singlet anomalous dimension
             :math:`\\gamma_{gq}^{(1)}(N)`
     """
-    S1 = harmonics.harmonic_S1(n)
-    S2 = harmonics.harmonic_S2(n)
-    Sp2p = harmonics.harmonic_S2(n / 2)
+    S1 = harmonics.S1(n)
+    S2 = harmonics.S2(n)
+    Sp2p = harmonics.S2(n / 2)
     # fmt: off
     ggq1_cfcf = (-8 + 2*n*(-12 + n*(-1 + n*(28 + n*(43 + 6*n*(5 + 2*n))))))/((-1 + n)*np.power(n,3)*np.power(1 + n,3)) - (4*(10 + n*(17 + n*(8 + 5*n)))*S1)/((-1 + n)*n*np.power(1 + n,2)) + (4*(2 + n + np.power(n,2))*np.power(S1,2))/(n*(-1 + np.power(n,2))) + (4*(2 + n + np.power(n,2))*S2)/(n*(-1 + np.power(n,2))) # pylint: disable=line-too-long
     ggq1_cfca = (-4*(144 + n*(432 + n*(-152 + n*(-1304 + n*(-1031 + n*(695 + n*(1678 + n*(1400 + n*(621 + 109*n))))))))))/(9.*np.power(n,3)*np.power(1 + n,3)*np.power(-2 + n + np.power(n,2),2)) + (4*(-12 + n*(-22 + 41*n + 17*np.power(n,3)))*S1)/(3.*np.power(-1 + n,2)*np.power(n,2)*(1 + n)) + ((8 + 4*n + 4*np.power(n,2))*np.power(S1,2))/(n - np.power(n,3)) + ((8 + 4*n + 4*np.power(n,2))*S2)/(n - np.power(n,3)) + (4*(2 + n + np.power(n,2))*Sp2p)/(n*(-1 + np.power(n,2))) # pylint: disable=line-too-long
@@ -214,15 +211,13 @@ def gamma_gg(n, nf):
             |NLO| gluon-gluon singlet anomalous dimension
             :math:`\\gamma_{gg}^{(1)}(N)`
     """
-    S1 = harmonics.harmonic_S1(n)
-    Sp1p = harmonics.harmonic_S1(n / 2)
-    Sp2p = harmonics.harmonic_S2(n / 2)
-    Sp3p = harmonics.harmonic_S3(n / 2)
-    g3n = harmonics.mellin_g3(n)
-    zeta2 = harmonics.zeta2
-    zeta3 = harmonics.zeta3
+    S1 = harmonics.S1(n)
+    Sp1p = harmonics.S1(n / 2)
+    Sp2p = harmonics.S2(n / 2)
+    Sp3p = harmonics.S3(n / 2)
+    g3n = harmonics.g_functions.mellin_g3(n)
     # fmt: off
-    ggg1_caca = 16*g3n - (2*(576 + n*(1488 + n*(560 + n*(-1248 + n*(-1384 + n*(1663 + n*(4514 + n*(4744 + n*(3030 + n*(1225 + 48*n*(7 + n))))))))))))/(9.*np.power(-1 + n,2)*np.power(n,3)*np.power(1 + n,3)*np.power(2 + n,3)) + S1*(29.77777777777778 + 16/np.power(-1 + n,2) + 16/np.power(1 + n,2) - 16/np.power(2 + n,2) - 8*Sp2p) + (16*(1 + n + np.power(n,2))*Sp2p)/(n*(1 + n)*(-2 + n + np.power(n,2))) - 2*Sp3p - 10*zeta3 + zeta2*(-16*S1 + 16*Sp1p + 16*(-(1/n) + np.log(2))) # pylint: disable=line-too-long
+    ggg1_caca = 16*g3n - (2*(576 + n*(1488 + n*(560 + n*(-1248 + n*(-1384 + n*(1663 + n*(4514 + n*(4744 + n*(3030 + n*(1225 + 48*n*(7 + n))))))))))))/(9.*np.power(-1 + n,2)*np.power(n,3)*np.power(1 + n,3)*np.power(2 + n,3)) + S1*(29.77777777777778 + 16/np.power(-1 + n,2) + 16/np.power(1 + n,2) - 16/np.power(2 + n,2) - 8*Sp2p) + (16*(1 + n + np.power(n,2))*Sp2p)/(n*(1 + n)*(-2 + n + np.power(n,2))) - 2*Sp3p - 10*zeta3 + zeta2*(-16*S1 + 16*Sp1p + 16*(-(1/n) + log2)) # pylint: disable=line-too-long
     ggg1_canf = (8*(6 + n*(1 + n)*(28 + n*(1 + n)*(13 + 3*n*(1 + n)))))/(9.*np.power(n,2)*np.power(1 + n,2)*(-2 + n + np.power(n,2))) - (40*S1)/9. # pylint: disable=line-too-long
     ggg1_cfnf = (2*(-8 + n*(-8 + n*(-10 + n*(-22 + n*(-3 + n*(6 + n*(8 + n*(4 + n)))))))))/(np.power(n,3)*np.power(1 + n,3)*(-2 + n + np.power(n,2))) # pylint: disable=line-too-long
     # fmt: on
