+++ src/eko/anomalous_dimensions/as3.py
@@ -9,11 +9,7 @@ Note that the QCD colour factors have been hard-wired in the parametrizations.
 import numba as nb
 import numpy as np
 
-from . import harmonics
-
-# Global variables
-zeta2 = harmonics.zeta2
-zeta3 = harmonics.zeta3
+from ..harmonics.constants import zeta2, zeta3
 
 
 @nb.njit(cache=True)
