+++ src/eko/harmonics/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+"""
+This module contains some harmonics sum.
+Defintion are coming from :cite:`MuselliPhD,Bl_mlein_2000,Blumlein:2009ta`
+"""
+
+from .w1 import S1, Sm1
+from .w2 import S2, Sm2
+from .w3 import S3, S21, S2m1, Sm2m1, Sm3, Sm21
+from .w4 import S4, S31, S211, Sm4, Sm22, Sm31, Sm211
+from .w5 import S5, S23, S41, S221, S311, S2111, S2m3, S21m2, Sm5, Sm23, Sm221, Sm2111
