+++ src/eko/harmonics/constants.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+"""
+Zeta function and other constants.
+"""
+import numpy as np
+from scipy.special import zeta
+
+zeta2 = zeta(2)
+r""":math:`\zeta(2)`"""
+
+zeta3 = zeta(3)
+r""":math:`\zeta(3)`"""
+
+zeta4 = zeta(4)
+r""":math:`\zeta(4)`"""
+
+zeta5 = zeta(5)
+r""":math:`\zeta(5)`"""
+
+
+log2 = np.log(2)
+r""":math:`\ln(2)`"""
+
+li4half = 0.517479
+""":math:`Li_{4}(1/2)`"""
